--- a/bus.go
+++ b/bus.go
@@ -57,7 +57,9 @@ func (b *messageBus) Subscribe(topic string, fn interface{}) error {
 
 	go func() {
 		for args := range h.queue {
-			h.callback.Call(args)
+			go func(targs []reflect.Value) {
+				h.callback.Call(targs)
+			}(args)
 		}
 	}()