process msg asynchronously to improve performance.