@vardius 2020-06-30
I merged it but have to think about it if this is an improvement or not, it increases allocations number
plus if the number of subscribers is high it actually slows down

before
```
BenchmarkPublish-4                   	 5214352	       226 ns/op	       0 B/op	       0 allocs/op
BenchmarkSubscribe-4                 	  632084	      3319 ns/op	     730 B/op	       5 allocs/op
Benchmark1Subscriber1Topic-4         	 5545311	       225 ns/op	       0 B/op	       0 allocs/op
Benchmark1Subscriber100Topics-4      	 2261838	       808 ns/op	       0 B/op	       0 allocs/op
Benchmark100Subscribers1Topic-4      	   61615	     19539 ns/op	       0 B/op	       0 allocs/op
Benchmark100Subscribers100Topics-4   	   60561	     34031 ns/op	       0 B/op	       0 allocs/op
```

after
```
BenchmarkPublish-4                   	  632365	      1935 ns/op	       1 B/op	       0 allocs/op
BenchmarkSubscribe-4                 	 1000000	      3950 ns/op	     726 B/op	       4 allocs/op
Benchmark1Subscriber1Topic-4         	  628796	      1835 ns/op	      11 B/op	       0 allocs/op
Benchmark1Subscriber100Topics-4      	  621848	      1902 ns/op	       0 B/op	       0 allocs/op
Benchmark100Subscribers1Topic-4      	    6956	    238003 ns/op	   12639 B/op	      28 allocs/op
Benchmark100Subscribers100Topics-4   	    6709	    298301 ns/op	       0 B/op	       0 allocs/op
```